+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+-   Move Agent Tasks under the Agents page
+-   Add a Plugin Tasks view under the Plugins page
+-   Add a Plugin Tasks view under the Plugin Execution page
 
 ## [2.1.1] - 2023-03-29